% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaSouce.R
\name{areaSource}
\alias{areaSource}
\title{distribution of emissions by region}
\format{a raster}
\usage{
areaSource(s, r, grid = NA, verbose = T)
}
\arguments{
\item{s}{input shape object}

\item{r}{input raster object}

\item{grid}{grid with the output format}

\item{verbose}{display adicional data}
}
\description{
Create the spetial distribution by a raster kasked by shape/model grid information.
}
\examples{
\dontrun{
# Do not run

shape  <- readOGR(paste(system.file("extdata", package = "EmissV"),"/BR.shp",sep=""),verbose = F)
shape  <- shape[22,1] # subset for Sao Paulo - BR
raster <- raster(paste(system.file("extdata", package = "EmissV"),"/sample.tiff",sep=""))
grid   <- gridInfo(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep=""))
SP     <- areaSource(shape,raster,grid)
spplot(SP, scales = list(draw=TRUE), xlab="Lat", ylab="Lon",main="Sao Paulo Metropolitan Area")

}
}
