% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmissV-package.R
\docType{package}
\name{EmissV-package}
\alias{EmissV}
\alias{EmissV-package}
\title{EmissV: Tools for Create Emissions for Air Quality Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Processing tools to create emissions for use in numerical air quality models. Emissions can be calculated both using emission factors and activity data (Schuch et al 2018) \doi{10.21105/joss.00662} or using pollutant inventories (Schuch et al., 2018) \doi{10.30564/jasr.v1i1.347}. Functions to process individual point emissions, line emissions and area emissions of pollutants are available as well as methods to incorporate alternative data for Spatial distribution of emissions such as satellite images (Gavidia-Calderon et. al, 2018) \doi{10.1016/j.atmosenv.2018.09.026} or openstreetmap data (Andrade et al, 2015) \doi{10.3389/fenvs.2015.00009}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://atmoschem.github.io/EmissV/}
  \item Report bugs at \url{https://github.com/atmoschem/EmissV/issues}
}

}
\author{
\strong{Maintainer}: Daniel Schuch \email{underschuch@gmail.com} (\href{https://orcid.org/0000-0001-5977-4519}{ORCID})

Authors:
\itemize{
  \item Sergio Ibarra-Espinosa \email{sergio.ibarra@usp.br} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})
}

}
\keyword{internal}
