% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newgrid.R
\name{newGrid}
\alias{newGrid}
\title{Create a grid set-up based on a NetCDF file}
\usage{
newGrid(file, levels = 1, verbose = T)
}
\arguments{
\item{file}{file name/path to a wrfinput of wrfchemi file}

\item{levels}{number of levels}

\item{verbose}{display adicional information}
}
\description{
Return a list contains information of a regular grid
}
\note{
its works with some wrf files (inicial condictions and emission) for now.
}
\examples{
\dontrun{
# Do not run
grid  <- newGrid(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep=""))
grid
}
}
