% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointSource.R
\name{pointSource}
\alias{pointSource}
\title{Emissions from point sources}
\usage{
pointSource(p = list(lat = -24.67123, lon = -47.26636, z = 100, e = 666),
  grid, verbose = T)
}
\arguments{
\item{p}{list of points}

\item{grid}{grid object with the grid information}

\item{verbose}{display additional information}
}
\description{
Transform a set of points into a grinded output
}
\seealso{
\code{\link{gridInfo}} and \code{\link{rasterSource}}
}
