% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointtogrid.R
\name{pointToGrid}
\alias{pointToGrid}
\title{create a point source}
\usage{
pointToGrid(p = list(x = -24.67123, y = -47.26636), grid, verbose = T)
}
\arguments{
\item{p}{list of points}

\item{grid}{grid object with the grid information}

\item{verbose}{display adicional information}
}
\description{
Transform a set of points into a grinded output
}
\seealso{
\code{\link{newGrid}} and \code{\link{rasterToGrid}}
}
