% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertogrid.R
\name{rasterToGrid}
\alias{rasterToGrid}
\title{Transform a raster into a grinded output}
\source{
exemple data from image of persistent lights of the Defense Meteorological Satellite Program (DMSP) \url{https://pt.wikipedia.org/wiki/Defense_Meteorological_Satellite_Program}
}
\usage{
rasterToGrid(r, grid, verbose = T)
}
\arguments{
\item{r}{input raster object}

\item{grid}{grid object with the grid information}

\item{verbose}{display adicional information}
}
\description{
Returns a matrix
}
\examples{
\dontrun{
# Do not run
grid  <- newGrid(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d01",sep=""))
x     <- raster(paste(system.file("extdata", package = "EmissV"),"/sample.tiff",sep=""))
test  <- rasterToGrid(x,grid)
image(test)
title("Persistent Nocturnal Lights from DMSP")
}
}
\seealso{
\code{\link{newGrid}} and \code{\link{shapeToGrid}}
}
