% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetogrid.R
\name{shapeToGrid}
\alias{shapeToGrid}
\title{Transform a shape into a grinded output}
\usage{
shapeToGrid(s, grid, verbose = T)
}
\arguments{
\item{s}{input shape object}

\item{grid}{grid object with the grid information}

\item{verbose}{display adicional information}
}
\description{
under development (test)
}
\examples{
\dontrun{
# Do not run
grid  <- newGrid(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep=""))
d3 <- data.frame(x = as.numeric(grid$Lon),y = as.numeric(grid$Lat))
d3 <- st_as_sf(d3, coords = c("x","y"))
st_crs(d3) <- st_crs(4326)

streets <- readRDS("D:/Downloads/streets.rds")
streets <- streets[streets$highway != "residential", ]

shapeToGrid(streets,d3)

}
}
\seealso{
\code{\link{newGrid}} and \code{\link{rasterToGrid}}
}
