% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalEmission.R
\name{totalEmission}
\alias{totalEmission}
\title{Calculate total emissions by territory (city, state, country, etc) by pollutant}
\format{Return the total emission (or a list) with the total emission (by day)}
\usage{
totalEmission(v, ef, pol = "CO", df = NA, verbose = T)
}
\arguments{
\item{v}{dataframe with the vehicle data}

\item{ef}{emission factors}

\item{pol}{pollutant name in ef}

\item{verbose}{display adicional information}
}
\description{
caculate the total emission with:

  Emission = sum( No_vehicles(n) * Km_day_use(n) * Emission_Factory(n,pollutant) (*Deterioration_Factors(n)) )

where n is the type of the veicle
}
\note{
its works with some wrf files (inicial condictions and emission) for now.
}
\examples{
\dontrun{
# Do not run

# FAZER !!

}
}
\seealso{
\code{\link{territory}}
}
