% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaSouce.R
\name{areaSource}
\alias{areaSource}
\title{Distribution of emissions by area}
\format{a raster}
\usage{
areaSource(s, r, grid = NA, name = "", as_frac = F, verbose = T)
}
\arguments{
\item{s}{input shape object}

\item{r}{input raster object}

\item{grid}{grid with the output format}

\item{name}{area name}

\item{as_frac}{return a fraction instead of a raster}

\item{verbose}{display additional data}
}
\description{
Calculate the spatial distribution by a raster kasked by shape/model grid information.
}
\examples{
\dontrun{
# Do not run

shape  <- raster::shapefile(paste(system.file("extdata", package = "EmissV"),
                            "/BR.shp",sep=""),verbose = F)
shape  <- shape[22,1] # subset for Sao Paulo - BR
raster <- raster::raster(paste(system.file("extdata", package = "EmissV"),
                         "/sample.tiff",sep=""))
grid   <- gridInfo(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep=""))
SP     <- areaSource(shape,raster,grid,name = "SPMA")
sp::spplot(SP, scales = list(draw=TRUE), xlab="Lat", ylab="Lon",
           main="Sao Paulo Metropolitan Area")

}
}
