% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emission.R
\name{emission}
\alias{emission}
\title{Emissions to atmospheric models}
\format{a matrix of emission}
\usage{
emission(total, pol, territorys, grid, mm = 1, aerosol = F, verbose = T)
}
\arguments{
\item{total}{list of total emission}

\item{pol}{pollutant name}

\item{territorys}{list of territory outputs}

\item{grid}{grid information}

\item{mm}{molar mass}

\item{aerosol}{TRUE for aerosols and FALSE (defoult) for gazes}

\item{verbose}{display adicional information}
}
\description{
Merge terrytory data with total emissions and convert to model output
}
\examples{
\dontrun{
# Do not run

}
}
\seealso{
\code{\link{totalEmission}} and \code{\link{territory}}
}
