% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emission.R
\name{emission}
\alias{emission}
\title{Emissions to atmospheric models}
\format{matrix of emission}
\usage{
emission(total, pol, territorys, grid, mm = 1, aerosol = F, verbose = T)
}
\arguments{
\item{total}{list of total emission}

\item{pol}{pollutant name}

\item{territorys}{list of territory outputs}

\item{grid}{grid information}

\item{mm}{pollutant molar mass}

\item{aerosol}{TRUE for aerosols and FALSE (defoult) for gazes}

\item{verbose}{display adicional information}
}
\description{
Combine territorial data and total emissions to model output
}
\note{
Is a god practice use the set_units(fe,your_unity), where fe is your emission factory and your_unity is usually g/km on your emission factory

the list of territorys must be in the same order as defined in vehicles and total emission.
}
\examples{
\dontrun{
# Do not run
veiculos <- vehicles(total_v = c(25141442, 5736428, 9147282),
                     territory_name = c("SP", "RJ", "MG"),
                     distribution = c( 0.4253, 0.0320, 0.3602, 0.0260,
                                       0.0290, 0.0008, 0.1181, 0.0086),
                     category =  c("LDV_E25","LDV_E100","LDV_F","TRUCKS_B5",
                                   "CBUS_B5","MBUS_B5","MOTO_E25","MOTO_F"),
                     type = c("LDV", "LDV", "LDV","TRUCKS",
                              "BUS","BUS","MOTO", "MOTO"),
                     fuel = c("E25", "E100", "FLEX","B5",
                              "B5","B5","E25", "FLEX"))

EmissionFactors <- as.data.frame.matrix(matrix(NA,ncol = 1,nrow = 8))
names(EmissionFactors) <- c("CO")
EmissionFactors["CO"]  <- set_units(rep(0.1,8),g/km)

TOTAL  <- totalEmission(veiculos,EmissionFactors,pol = c("CO"),verbose = T)

grid   <- newGrid(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01"))
shape  <- readOGR(paste0(system.file("extdata", package = "EmissV"),"/BR.shp"))
raster <- raster(paste0(system.file("extdata", package = "EmissV"),"/sample.tiff"))

SP     <- territory(shape[22,1],raster,grid)
RJ     <- territory(shape[17,1],raster,grid)
MG     <- territory(shape[12,1],raster,grid)

e_CO   <- emission(TOTAL,"CO",list(SP = SP, RJ = RJ, MG = MG),grid)
}
}
\seealso{
\code{\link{totalEmission}} and \code{\link{territory}}
}
