% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emission.R
\name{emission}
\alias{emission}
\title{Emissions to atmospheric models}
\format{matrix of emission}
\usage{
emission(total, pol, area, grid, mm = 1, aerosol = F, verbose = T)
}
\arguments{
\item{total}{list of total emission}

\item{pol}{pollutant name}

\item{area}{list of area sources or matrix with a spatial distribution}

\item{grid}{grid information}

\item{mm}{pollutant molar mass}

\item{aerosol}{TRUE for aerosols and FALSE (defoult) for gazes}

\item{verbose}{display adicional information}
}
\description{
Combine area sourses and total emissions to model output
}
\note{
Is a god practice use the set_units(fe,your_unity), where fe is your emission factory and your_unity is usually g/km on your emission factory

for matrices just the first value of total is used

the list of area must be in the same order as defined in vehicles and total emission.
}
\examples{
\dontrun{
# Do not run

# DETRAN 2016 data and SP vahicle distribution
veiculos <- vehicles(total_v = c(27332101, 6377484, 10277988),
                     area_name = c("SP", "RJ", "MG"),
                     distribution = c( 0.4253, 0.0320, 0.3602, 0.0260,
                                       0.0290, 0.0008, 0.1181, 0.0086),
                     category =  c("LDV_E25","LDV_E100","LDV_F","TRUCKS_B5",
                                   "CBUS_B5","MBUS_B5","MOTO_E25","MOTO_F"),
                     type = c("LDV", "LDV", "LDV","TRUCKS",
                              "BUS","BUS","MOTO", "MOTO"),
                     fuel = c("E25", "E100", "FLEX","B5",
                              "B5","B5","E25", "FLEX"),
                     vnames = c("Light duty Vehicles Gasohol","Light Duty Vehicles Ethanol",
                                "Light Duty Vehicles Flex","Diesel trucks","Diesel urban busses",
                                "Diesel intercity busses","Gasohol motorcycles","Flex motorcycles"))

# values calculated from CETESB 2015 with
# weighted.mean( emissions by type and year, DETRAN frota by type and year)
# for Sao Paulo
EmissionFactors <- as.data.frame.matrix(matrix(NA,ncol = 1,nrow = 8))
names(EmissionFactors) <- c("CO")
EmissionFactors$CO <- set_units(c(1.75,10.04,0.39,0.45,0.77,1.48,1.61,0.75),g/km)
rownames(EmissionFactors) <- c("Light duty Vehicles Gasohol","Light Duty Vehicles Ethanol",
                               "Light Duty Vehicles Flex","Diesel trucks","Diesel urban busses",
                               "Diesel intercity busses","Gasohol motorcycles","Flex motorcycles")

TOTAL  <- totalEmission(veiculos,EmissionFactors,pol = c("CO"),verbose = T)

grid   <- gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01"))
shape  <- readOGR(paste0(system.file("extdata", package = "EmissV"),"/BR.shp"),verbose = F)
raster <- raster(paste0(system.file("extdata", package = "EmissV"),"/sample.tiff"))

SP     <- areaSource(shape[22,1],raster,grid)
RJ     <- areaSource(shape[17,1],raster,grid)
MG     <- areaSource(shape[12,1],raster,grid)

e_CO   <- emission(TOTAL,"CO",list(SP = SP, RJ = RJ, MG = MG),grid,mm=28)
}
}
\seealso{
\code{\link{totalEmission}} and \code{\link{areaSource}}
}
