% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridinfo.R
\name{gridInfo}
\alias{gridInfo}
\title{read grid information from a NetCDF file}
\usage{
gridInfo(file = file.choose(), levels = 1, verbose = T)
}
\arguments{
\item{file}{file name/path to a wrfinput of wrfchemi file}

\item{levels}{number of levels}

\item{verbose}{display adicional information}
}
\description{
Return a list contains information of a regular grid / domain
}
\note{
its works with some wrf files (inicial condictions and emission) for now.
}
\examples{
\dontrun{
# Do not run
grid  <- gridInfo(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep=""))
grid
}
}
