% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfil.R
\docType{data}
\name{perfil}
\alias{perfil}
\title{Temporal profile for veicular emissions}
\format{A list of data frames by hour and weekday.}
\usage{
data(perfil)
}
\description{
List with data frames with temporal profile based on traffic count for Brazilian city Sao Paulo.

LDV is for light vehicles and HDV for intencity of heavy vehicles.
}
\note{
The profile is normalized by days (but is balanced for a complete week) so diary emission x profile = hourly emission.
}
\examples{
\dontrun{
# Do not run

# load the data
data(perfil)

# function to simple view
plot.perfil <- function(per = perfil$LDV, text="", color = "#0000FFBB"){
plot(per[,1],ty = "l", ylim = range(per),axe = F,xlab = "hour",ylab = "Intensity",main = text,col=color)
for(i in 2:7){
lines(per[,i],col = color)
}
for(i in 1:7){
points(per[,i],col = "black", pch = 20)
}
axis(1,at=0.5+c(0,6,12,18,24),labels = c("00:00","06:00","12:00","18:00","00:00"))
axis(2)
box()
}

# view all profiles in perfil data
for(i in 1:length(names(perfil))){
print(paste("profile",i,names(perfil)[i]))
plot.perfil(perfil[[i]],names(perfil)[i])
}
}
}
\keyword{datasets}
