% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumerise.R
\name{plumeRise}
\alias{plumeRise}
\title{Calculate plume rise from Brigs (1975) formulation.}
\format{numeric value in (units) meters}
\usage{
plumeRise(df, imax = 100, errmax = 0.1, verbose = T)
}
\arguments{
\item{df}{data.frame with micrometeorological and source data}

\item{verbose}{display additional information}
}
\description{
Return plume rise height
}
\examples{
\dontrun{
# Do not run

chimney <- matrix(c(1,1,1,1,1,1,1,1,1,1),ncol = 10, byrow = T)
chimney<- data.frame(chimney)
names(chimney) <- c("z","r","Ve","Te","h","ws","Temp","Ustar","Wstar","L")

chimay <- plumeRise(chimney)

}
}
