% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumerise.R
\name{plumeRise}
\alias{plumeRise}
\title{Calculate plume rise from Brigs (1975) formulation.}
\format{a numeric value in (units) meters}
\usage{
plumeRise(verbose = T)
}
\arguments{
\item{verbose}{display adicional information}
}
\description{
return plume rise
}
\examples{
\dontrun{
# Do not run


}
}
