% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointtogrid.R
\name{pointToGrid}
\alias{pointToGrid}
\title{create a point source}
\usage{
pointToGrid(p, grid, verbose = T)
}
\arguments{
\item{p}{input point object}

\item{grid}{grid object with the grid information}

\item{verbose}{display adicional information}
}
\description{
under development (test)
}
\seealso{
\code{\link{newGrid}} and \code{\link{rasterToGrid}}
}
