% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSource.R
\name{rasterSource}
\alias{rasterSource}
\title{Distribution of emissions by a georeferenced image}
\source{
Exemple data from image of persistent lights of the Defense Meteorological Satellite Program (DMSP) \url{https://pt.wikipedia.org/wiki/Defense_Meteorological_Satellite_Program}
}
\usage{
rasterSource(r, grid, verbose = T)
}
\arguments{
\item{r}{input raster object}

\item{grid}{grid object with the grid information}

\item{verbose}{display additional information}
}
\description{
Returns a matrix
}
\examples{
\dontrun{
# Do not run
grid  <- gridInfo(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d01",sep=""))
x     <- raster(paste(system.file("extdata", package = "EmissV"),"/sample.tiff",sep=""))
test  <- rasterSource(x,grid)
image(test)
title("Persistent Nocturnal Lights from DMSP")
}
}
\seealso{
\code{\link{gridInfo}} and \code{\link{lineSource}}
}
