% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapetogrid.R
\name{shapeToGrid}
\alias{shapeToGrid}
\title{Transform a shape into a grinded output}
\usage{
shapeToGrid(s, grid, verbose = T)
}
\arguments{
\item{s}{input shape object}

\item{grid}{grid object with the grid information}

\item{verbose}{display adicional information}
}
\description{
under development
}
\seealso{
\code{\link{newGrid}} and \code{\link{rasterToGrid}}
}
