% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territory.R
\name{territory}
\alias{territory}
\title{Subset a raster with a shape (and grid)}
\usage{
territory(s, r, grid = NA, verbose = T)
}
\arguments{
\item{s}{input shape object}

\item{r}{input raster object}

\item{grid}{grid with the output format}

\item{verbose}{display adicional data}
}
\description{
Create a subset of a raster by a shape and return a new masked raster. If grid argument is used return in grid format and the fraction within the grid.
}
\examples{
\dontrun{
# Do not run

shape  <- read_sf(paste(system.file("extdata", package = "EmissV"),"/BR.shp",sep=""),verbose = F)
shape  <- shape[22,1] # subset for Sao Paulo - BR
raster <- raster(paste(system.file("extdata", package = "EmissV"),"/sample.tiff",sep=""))
grid   <- newGrid(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep=""))
SP     <- territory(shape,raster,grid)
plot(SP,main = "Sao Paulo Metropolitan Area", xlab="Lon",ylab="Lat",
     col = heat.colors(12, alpha = 1))
}
}
