% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalEmission.R
\name{totalEmission}
\alias{totalEmission}
\title{Calculate total emissions}
\format{Return a list with the daily total emission by territory.}
\usage{
totalEmission(v, ef, pol, verbose = T)
}
\arguments{
\item{v}{dataframe with the vehicle data}

\item{ef}{emission factors}

\item{pol}{pollutant name in ef}

\item{verbose}{display adicional information}
}
\description{
caculate the total emission with:

  Emission = sum( Vehicles(n) * Km_day_use(n) * Emission_Factory(n,pollutant) (*Deterioration_Factors(n)) )

where n is the type of the veicle
}
\note{
the units can be used
}
\examples{
\dontrun{
# Do not run

veiculos <- vehicles(total_v = c(25141442, 5736428, 9147282, 6523727, 4312896),
                     territory_name = c("SP", "RJ", "MG", "PR", "SC"),
                     distribution = c( 0.4253, 0.0320, 0.3602, 0.0260, 0.0290, 0.0008, 0.1181, 0.0086),
                     category =  c("LDV_E25","LDV_E100","LDV_F","TRUCKS_B5","CBUS_B5","MBUS_B5","MOTO_E25","MOTO_F"),
                     type = c("LDV", "LDV", "LDV","TRUCKS","BUS","BUS","MOTO", "MOTO"),
                     fuel = c("E25", "E100", "FLEX","B5","B5","B5","E25", "FLEX"),
                     vnames = c("Light duty Vehicles Gasohol","Light Duty Vehicles Ethanol","Light Duty Vehicles Flex","Diesel trucks",
                                "Diesel urban busses","Diesel intercity busses","Gasohol motorcycles","Flex motorcycles"))

EmissionFactors <- as.data.frame.matrix(matrix(NA,ncol = 2,nrow = 8))
rownames(EmissionFactors) <- c("Light duty Vehicles Gasohol","Light Duty Vehicles Ethanol","Light Duty Vehicles Flex","Diesel trucks",
                               "Diesel urban busses","Diesel intercity busses","Gasohol motorcycles","Flex motorcycles")
names(EmissionFactors) <- c("CO","HC")
# set the correct units
EmissionFactors$CO <- set_units(rep(0.1,8),g/km)
EmissionFactors$HC <- set_units(rep(0.15,8),g/km)

TOTAL <- totalEmission(veiculos,EmissionFactors,pol = c("CO","HC"))

}
}
\seealso{
\code{\link{territory}}
}
