% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalvoc.R
\name{totalVOC}
\alias{totalVOC}
\title{Calculate Total VOCs emissions}
\format{Return a list with the daily total emission by territory.}
\usage{
totalVOC(v, ef, pol, verbose = T)
}
\arguments{
\item{v}{data frame with the vehicle data}

\item{ef}{emission factors}

\item{pol}{pollutant name in ef}

\item{verbose}{display additional information}
}
\description{
Calculate Volatile Organic Compounds (COVs) emited by the process of exhaust (through the exhaust pipe), liquid (carter and evaporative) and  vapor (fuel transfer operations).

 Avaliable VOCs are: eth, hc3, hc5, hc8, ol2, olt, oli, iso, tol, xyl, ket, ch3oh and ald
}
\note{
The same ef can be used to totalEmission and voc
}
\examples{
\dontrun{
# Do not run

# DETRAN 2016 data and SP vahicle distribution
veiculos <- vehicles(total_v = c(27332101, 6377484, 10277988, 7140439, 4772160),
                     area_name = c("SP", "RJ", "MG", "PR", "SC"),
                     distribution = c( 0.4253, 0.0320, 0.3602, 0.0260,
                                       0.0290, 0.0008, 0.1181, 0.0086),
                     category =  c("LDV_E25","LDV_E100","LDV_F","TRUCKS_B5",
                                   "CBUS_B5","MBUS_B5","MOTO_E25","MOTO_F"),
                     type = c("LDV", "LDV", "LDV","TRUCKS",
                              "BUS","BUS","MOTO", "MOTO"),
                     fuel = c("E25", "E100", "FLEX","B5",
                              "B5","B5","E25", "FLEX"),
                     vnames = c("Light duty Vehicles Gasohol","Light Duty Vehicles Ethanol",
                                "Light Duty Vehicles Flex","Diesel trucks","Diesel urban busses",
                                "Diesel intercity busses","Gasohol motorcycles","Flex motorcycles"))

voc = c("eth","hc3","hc5","hc8","ol2","olt","oli","iso","tol","xyl","ket","ch3oh","ald")

EF_voc <- as.data.frame(matrix(NA,ncol = 9,nrow = 8,byrow = T),
                        row.names = c("Light duty Vehicles Gasohol",
                                      "Light Duty Vehicles Ethanol",
                                      "Light Duty Vehicles Flex","Diesel trucks",
                                      "Diesel urban busses","Diesel intercity busses",
                                      "Gasohol motorcycles","Flex motorcycles"))
names(EF_voc) <-  c("voc_vap_g","voc_vap_e","voc_vap_d",
                    "voc_liq_g","voc_liq_e","voc_liq_d",
                    "voc_exa_g","voc_exa_e","voc_exa_d")

EF_voc["voc_vap_g"]  <- c(0.230,0.00,0.120,0.00,0.00,0.00,0.00,0.00)
EF_voc["voc_vap_e"]  <- c(0.000,0.25,0.120,0.00,0.00,0.00,0.00,0.00)
EF_voc["voc_vap_d"]  <- c(0.000,0.00,0.000,0.00,0.00,0.00,0.00,0.00)
EF_voc["voc_liq_g"]  <- c(2.000,0.00,0.875,0.00,0.00,0.00,1.20,0.00)
EF_voc["voc_liq_e"]  <- c(0.000,1.50,0.875,0.00,0.00,0.00,0.00,1.20)
EF_voc["voc_liq_d"]  <- c(0.000,0.00,0.000,0.00,0.00,0.00,0.00,0.00)
EF_voc["voc_exa_g"]  <- c(0.425,0.00,0.217,0.00,0.00,0.00,1.08,0.00)
EF_voc["voc_exa_e"]  <- c(0.000,1.30,0.217,0.00,0.00,0.00,0.00,1.08)
EF_voc["voc_exa_d"]  <- c(0.000,0.00,0.000,2.05,0.00,0.00,0.00,0.00)

print(EF_voc)

VOC <- totalVOC(veiculos,EF_voc,pol = voc[10])
}
}
\seealso{
\code{\link{totalEmission}} and \code{\link{vehicles}}
}
